# apps/api/pkg/entities/user.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.8
// 	protoc        v3.19.6
// source: pkg/entities/user.proto

package entities

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type User struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Email         string                 `protobuf:"bytes,1,opt,name=email,proto3" json:"email,omitempty"`
	Username      string                 `protobuf:"bytes,2,opt,name=username,proto3" json:"username,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *User) Reset() {
	*x = User{}
	mi := &file_pkg_entities_user_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *User) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*User) ProtoMessage() {}

func (x *User) ProtoReflect() protoreflect.Message {
	mi := &file_pkg_entities_user_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use User.ProtoReflect.Descriptor instead.
func (*User) Descriptor() ([]byte, []int) {
	return file_pkg_entities_user_proto_rawDescGZIP(), []int{0}
}

func (x *User) GetEmail() string {
	if x != nil {
		return x.Email
	}
	return ""
}

func (x *User) GetUsername() string {
	if x != nil {
		return x.Username
	}
	return ""
}

var File_pkg_entities_user_proto protoreflect.FileDescriptor

const file_pkg_entities_user_proto_rawDesc = "" +
	"\n" +
	"\x17pkg/entities/user.proto\x12\bentities\"8\n" +
	"\x04User\x12\x14\n" +
	"\x05email\x18\x01 \x01(\tR\x05email\x12\x1a\n" +
	"\busername\x18\x02 \x01(\tR\busernameB\x1bZ\x19ekel-backend/pkg/entitiesb\x06proto3"

var (
	file_pkg_entities_user_proto_rawDescOnce sync.Once
	file_pkg_entities_user_proto_rawDescData []byte
)

func file_pkg_entities_user_proto_rawDescGZIP() []byte {
	file_pkg_entities_user_proto_rawDescOnce.Do(func() {
		file_pkg_entities_user_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_pkg_entities_user_proto_rawDesc), len(file_pkg_entities_user_proto_rawDesc)))
	})
	return file_pkg_entities_user_proto_rawDescData
}

var file_pkg_entities_user_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
var file_pkg_entities_user_proto_goTypes = []any{
	(*User)(nil), // 0: entities.User
}
var file_pkg_entities_user_proto_depIdxs = []int32{
	0, // [0:0] is the sub-list for method output_type
	0, // [0:0] is the sub-list for method input_type
	0, // [0:0] is the sub-list for extension type_name
	0, // [0:0] is the sub-list for extension extendee
	0, // [0:0] is the sub-list for field type_name
}

func init() { file_pkg_entities_user_proto_init() }
func file_pkg_entities_user_proto_init() {
	if File_pkg_entities_user_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_pkg_entities_user_proto_rawDesc), len(file_pkg_entities_user_proto_rawDesc)),
			NumEnums:      0,
			NumMessages:   1,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_pkg_entities_user_proto_goTypes,
		DependencyIndexes: file_pkg_entities_user_proto_depIdxs,
		MessageInfos:      file_pkg_entities_user_proto_msgTypes,
	}.Build()
	File_pkg_entities_user_proto = out.File
	file_pkg_entities_user_proto_goTypes = nil
	file_pkg_entities_user_proto_depIdxs = nil
}